+++ scriptabit/utility_functions.py
@@ -88,6 +88,15 @@ class UtilityFunctions(object):
 
         return parser
 
+    @property
+    def dry_run(self):
+        """ Indicates whether this is a dry run or not.
+
+        Returns:
+            bool: True if this is a dry run, otherwise False.
+        """
+        return self.__config.dry_run
+
     def run(self):
         """Runs the user-selected scriptabit utility functions"""
         if self.__config.test:
@@ -112,7 +121,7 @@ class UtilityFunctions(object):
     def set_health(self, hp):
         """Sets the user health to the specified value"""
         old_hp = self.__hs.get_stats()['hp']
-        new_hp = self.__hs.set_hp(hp)
+        new_hp = hp if self.dry_run else self.__hs.set_hp(hp)
         logging.getLogger(__name__).info(
             'HP changed from %f to %f',
             old_hp,
@@ -121,7 +130,7 @@ class UtilityFunctions(object):
     def set_xp(self, xp):
         """Sets the user experience points to the specified value"""
         old_xp = self.__hs.get_stats()['exp']
-        new_xp = self.__hs.set_exp(xp)
+        new_xp = xp if self.dry_run else self.__hs.set_exp(xp)
         logging.getLogger(__name__).info(
             'XP changed from %f to %f',
             old_xp,
@@ -130,7 +139,7 @@ class UtilityFunctions(object):
     def set_mana(self, mp):
         """Sets the user mana to the specified value"""
         old_mp = self.__hs.get_stats()['mp']
-        new_mp = self.__hs.set_mp(mp)
+        new_mp = mp if self.dry_run else self.__hs.set_mp(mp)
         logging.getLogger(__name__).info(
             'MP changed from %f to %f',
             old_mp,
@@ -139,7 +148,7 @@ class UtilityFunctions(object):
     def set_gold(self, gp):
         """Sets the user gold to the specified value"""
         old = self.__hs.get_stats()['gp']
-        new = self.__hs.set_gp(gp)
+        new = gp if self.dry_run else self.__hs.set_gp(gp)
         logging.getLogger(__name__).info(
             'Gold changed from %f to %f',
             old,