+++ CHANGELOG.rst
@@ -208,3 +208,4 @@ Changelog
 * Implemented full dry run mode support in CSV uploader.
 * Implemented full dry run mode support in pet care plugin.
 * Implemented full dry run mode support in trello sync plugin.
+* Implemented full dry run mode support in the utility functions.